% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/App_main.R
\name{NASSboard}
\alias{NASSboard}
\title{GUI for Interactive Mapping of NASS Census Data}
\usage{
NASSboard()
}
\description{
Running this function with no arguments will launch a general user interface for
mapping NASS data at the county level.
}
\examples{
NASSboard()

}
